Create version 4-7-15 html
I couldn't get the local web server to work, I look into it more later. In this html file, I added some buttons and the form for adding or editing a map. The problem is that I haven't figure out how to overwrite an element, so it kinda bunch together when you click the buttons. I will look more into that, and try to fix it.