Create NoServerConnectionVersion
So, I got the add form button to appear by itself. But there is a problem with the edit button. When the edit button is click first, followed by the add button, both of the forms will show. I will try to fix that.

Also, the data inside the NPC button and Location button have hover function to highlight the corresponding thing in the map. But right now, it can't highlight the stuff, it only shows alert because I haven't figured out how to do it yet.

Also, after clicking the confirm edit button in the edit form, it will show alerts of what you entered. So, all I need to do is to add those to the database, but I still don't know how to do the database stuff yet since I can't connect to the server.