Did some stuff
Fixed the Description, NPC, and Location buttons so they actually link
to the json file now.
Also, I was getting a weird issue with the web.config file not working
for me on my desktop, so I modified it. Finally, the json file now has
sample data for child maps